@@ -151,8 +151,8 @@
     "i18next-http-backend": "^1.4.0",
     "ini": "^3.0.0",
     "plist": "^3.0.5",
-    "react": "17",
-    "react-dom": "17",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-i18next": "^11.16.7",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
