@@ -1,7 +1,7 @@
 import { I18nextProvider, initReactI18next } from 'react-i18next'
 import HttpApi from 'i18next-http-backend'
 import React, { Suspense } from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import i18next from 'i18next'
 import { initGamepad } from './helpers/gamepad'
 
@@ -71,7 +71,11 @@ i18next
     ]
   })
 
-ReactDOM.render(
+const container = document.getElementById('root')
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+const root = createRoot(container!)
+
+root.render(
   <React.StrictMode>
     <I18nextProvider i18n={i18next}>
       <Suspense fallback={<UpdateComponentBase message="Loading" />}>
@@ -80,6 +84,5 @@ ReactDOM.render(
         </GlobalState>
       </Suspense>
     </I18nextProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 )
